backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2.2"
python-slugify = "^5.0.2"
djangorestframework = "^3.12.4"
django-cors-headers = "^3.7.0"
pytest = "^6.2.4"
pytest-django = "^4.3.0"
Pillow = "^8.2.0"
pytest-factoryboy = "^2.1.0"
pytest-cov = "^2.12.1"

[tool.poetry.dev-dependencies]
black = "^21.6b0"
isort = "^5.9.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"